+++ src/adc/api/_writer.py
@@ -28,7 +28,7 @@ class SplittableBatchWriter(BatchWriter):
     Ancestor for dataset batch writers.
     """
 
-    def __init__(self, split_names: List[str] = None, split_ratios: List[int] = None,
+    def __init__(self, split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
         Initializes the reader.
@@ -37,6 +37,8 @@ class SplittableBatchWriter(BatchWriter):
         :type split_names: list
         :param split_ratios: the integer ratios of the splits (must sum up to 100)
         :type split_ratios: list
+        :param split_group: the regular expression with a single group used for keeping items in the same split, e.g., for identifying the base name of a file or the ID
+        :type split_group: str
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -45,8 +47,9 @@ class SplittableBatchWriter(BatchWriter):
         super().__init__(logger_name=logger_name, logging_level=logging_level)
         self.split_names = None
         self.split_ratios = None
+        self.split_group = None
         self.splitter = None
-        seppl.io.init_splitting_params(self, split_names=split_names, split_ratios=split_ratios)
+        seppl.io.init_splitting_params(self, split_names=split_names, split_ratios=split_ratios, split_group=split_group)
 
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
@@ -99,7 +102,7 @@ class SplittableStreamWriter(StreamWriter):
     Ancestor for dataset stream writers.
     """
 
-    def __init__(self, split_names: List[str] = None, split_ratios: List[int] = None,
+    def __init__(self, split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
         Initializes the reader.
@@ -108,6 +111,8 @@ class SplittableStreamWriter(StreamWriter):
         :type split_names: list
         :param split_ratios: the integer ratios of the splits (must sum up to 100)
         :type split_ratios: list
+        :param split_group: the regular expression with a single group used for keeping items in the same split, e.g., for identifying the base name of a file or the ID
+        :type split_group: str
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -116,8 +121,9 @@ class SplittableStreamWriter(StreamWriter):
         super().__init__(logger_name=logger_name, logging_level=logging_level)
         self.split_names = None
         self.split_ratios = None
+        self.split_group = None
         self.splitter = None
-        seppl.io.init_splitting_params(self, split_names=split_names, split_ratios=split_ratios)
+        seppl.io.init_splitting_params(self, split_names=split_names, split_ratios=split_ratios, split_group=split_group)
 
     def _create_argparser(self) -> argparse.ArgumentParser:
         """